Add fuzzy diff for backdrop-filter-fixed-clip
This test is a follow up to
https://chromium-review.googlesource.com/c/chromium/src/+/6916525, which
fixed another test that also had a 1/255 diff in a color channel. This
diff is caused by differences in rendering pipelines and is not a code
issue.

This test already passes with SKIA, but fails with software renderer.
This CL fixes that.

Bug: 56387505
Change-Id: I86aa7e27cd2743ecc8bf66d429b5cd86b684efe6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6941602
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Claire Chambers <clchambers@microsoft.com>
Reviewed-by: Olga Gerchikov <gerchiko@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1521393}